refactor(email): déplacer le contenu dans src/content/emails/
Le fichier de contenu passe de l'infra vers src/content/emails/,
un dossier CMS-like versionné pour tous les contenus emails.
