+++ src/content/emails/onboarding-welcome.content.ts
